+++ R/getHistory.R
@@ -4,6 +4,7 @@
 #' 
 #' @template  server 
 #' @param  days Number of days to get history for defaults to 1L.
+#' @param filter optional string to filter the results against
 #' @template  dryRun 
 #' @template ellipsis
 #'
@@ -13,12 +14,13 @@
 #'
 #' @example /inst/examples/docs/getHistory.R
 
-getHistory <- function(server, days = 1L, dryRun = FALSE, ...){
+getHistory <- function(server, days = 1L, filter = NULL, dryRun = FALSE, ...){
   checkArgs(server, days, dryRun)
   body <- setNames(
-    list(days, "csv", "on"),
-    c("days", "f", "all")
+    list(days, "csv", "on", filter),
+    c("days", "f", "all", "filter")
   )
+  body[["filter"]]
   res <- sendQuery(url = server$url, path = WPTPaths$history, 
                    body = body, origin = "getHistory", dryRun = dryRun,
                    method = "GET", ...)